requirements.txt
requests>=2.31.0
Pillow>=10.0.0
pyobjc-framework-Cocoa>=9.0